+++ src/components/atoms/Toast.tsx
@@ -1,6 +1,13 @@
-import { createContext, useContext } from "react";
+import {
+  createContext,
+  useCallback,
+  useState,
+  useContext,
+  ReactNode,
+} from "react";
 import { createPortal } from "react-dom";
 import clsx from "clsx";
+import { v4 as uuid } from "uuid";
 import { Icon } from "./Icon";
 
 export function Toast() {
@@ -10,7 +17,7 @@ export function Toast() {
         <div
           className={clsx(
             "relative",
-            "flex items-center justify-between py-2.5 px-6 rounded-xl",
+            "flex items-center justify-between py-2.5 px-6 rounded-lg",
             " bg-orange m-0.5 text-white opacity-90",
             " w-80 md:w-auto",
             "transition-all duration-100 animate-moveDown"
@@ -38,6 +45,42 @@ type Message = {
 type SetMessageAction = (msg: Message[]) => void;
 
 const Context = createContext<SetMessageAction | undefined>(undefined);
+
+type ToastProviderProps = {
+  children: ReactNode;
+};
+
+export function ToastProvider({ children }: ToastProviderProps) {
+  const [messages, setMessages] = useState<
+    Array<{ id: string; message: string }>
+  >([]);
+
+  const setMessage = useCallback(
+    (message) => {
+      if (!message) return;
+
+      console.log(message);
+      const id = uuid();
+
+      setMessages((queue) => [...queue, { id, message }]);
+
+      setTimeout(() => {
+        setMessages((queue) => queue.filter((pair) => pair.id !== id));
+      }, 2000);
+    },
+    [setMessages]
+  );
+
+  return (
+    <Context.Provider value={setMessage}>
+      {children}
+
+      {messages.map(({ id }) => (
+        <Toast key={id} />
+      ))}
+    </Context.Provider>
+  );
+}
 
 export function useToast() {
   const context = useContext(Context);
