+++ src/layouts/Home.tsx
@@ -2,12 +2,13 @@ import { FormEvent, ChangeEvent } from "react";
 import { has } from "ramda";
 import { useNavigate } from "react-router-dom";
 
-import { Input, Button } from "@/components";
+import { Input, Button, useToast } from "@/components";
 import { API, Query, useDispatch, useSelector } from "@/logic";
 import { pickRandomIn, URLSearchParams } from "@/utils";
 import logo from "@/assets/images/logo.png";
 import logoWb from "@/assets/svgs/home-logo.svg";
 import TITLE from "@/assets/title.json";
+import { v4 as uuid } from "uuid";
 
 const title = pickRandomIn(TITLE);
 
@@ -15,6 +16,8 @@ export function Home() {
   const query = useSelector(Query.selectQuery);
   const dispatch = useDispatch();
   const navigate = useNavigate();
+  const setToast = useToast();
+  setToast([{ id: uuid(), message: "oh" }]);
 
   function onChange(event: ChangeEvent<HTMLInputElement>) {
     dispatch(Query.update(event.target.value));
