+++ src/logic/providers/index.tsx
@@ -1,3 +1,2 @@
 export * from "./geo";
 export * from "./system";
-export * from "./toast";
